yk: revised.

+++ infer/src/absint/AbstractDomain.ml
@@ -29,7 +29,7 @@ module type NoJoinForML = sig
 
   val leq : lhs:t -> rhs:t -> bool
   val similar : lhs:t -> rhs:t -> bool
-  val feature_vector : t -> float lazy_t list
+  val feature_vector : t -> int lazy_t list
 end
 
 module type S = sig
